修改返回类型，并简化一部分代码

+++ dorive-core/src/main/java/com/gitee/dorive/core/api/executor/EntityHandler.java
@@ -23,6 +23,6 @@ import java.util.List;
 
 public interface EntityHandler {
 
-    int handle(Context context, List<Object> entities);
+    long handle(Context context, List<Object> entities);
 
 }
